adr1/requirements.txt
aiosqlite==0.17.0
anyio==3.5.0
asgiref==3.5.0
asyncpg==0.25.0
click==8.1.2
fastapi==0.75.1
h11==0.13.0
idna==3.3
iso8601==1.0.2
psycopg2-binary==2.9.3
pydantic==1.9.0
pypika-tortoise==0.1.4
pytz==2022.1
PyYAML==6.0
sniffio==1.2.0
starlette==0.17.1
tortoise-orm==0.19.0
typing-extensions==4.1.1
uvicorn==0.17.6
